* tools/tests/8_funs/f8.c: Use types correctly.

+++ tools/tests/8_funs/f8.c
@@ -18,7 +18,7 @@ unsigned int foo() { return ~0L; }
 
 sl_def(t_main, void)
 {
-  unsigned long x = sl_funcall(,ulong,foo);
+  unsigned int x = sl_funcall(,ulong,foo);
   output_uint(~x, 1); output_char('\n', 1);
 }
 sl_enddef
